+++ Akane/Akane/Component/View/ComponentView.swift
@@ -9,10 +9,17 @@
 import Foundation
 import HasAssociatedObjects
 
-public protocol Displayable {
-    associatedtype Props
+public protocol Displayable : class {
+    associatedtype Parameters
 
-    func bindings(_ observer: ViewObserver, props: Props)
+    /**
+     Define the bindings between displayable and its params
+
+     - parameter observer:  The observer to use for defining  and registering
+     bindings
+     - parameter viewModel: `Parameters` associated to Displayable
+     */
+    func bindings(_ observer: ViewObserver, params: Parameters)
 }
 
 public protocol _AnyComponentDisplayable {
@@ -26,17 +33,7 @@ public typealias ComponentView = ComponentDisplayable
 ComponentView is used on an `UIView` in order to associate it to a 
 `ComponentViewModel` implementing its business logic.
 */
-public protocol ComponentDisplayable : class, HasAssociatedObjects, _AnyComponentDisplayable {
-    associatedtype ViewModelType: ComponentViewModel
-
-    /**
-    Define the bindings between the fields (IBOutlet) and the ComponentViewModel
-    
-    - parameter observer:  The observer to use for defining  and registering 
-    bindings
-    - parameter viewModel: The `ComponentViewModel` associated to the `UIView`
-    */
-    func bindings(_ observer: ViewObserver, viewModel: ViewModelType)
+public protocol ComponentDisplayable : Displayable, HasAssociatedObjects, _AnyComponentDisplayable where Parameters : ComponentViewModel {
 }
 
 /**
@@ -48,7 +45,7 @@ public protocol Wrapped {
 
 extension ComponentDisplayable where Self : UIView {
     public func _tryBindings(_ observer: ViewObserver, viewModel: Any) {
-        guard let viewModel = viewModel as? ViewModelType else {
+        guard let viewModel = viewModel as? Parameters else {
             return
         }